raseed-backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3

# Firebase
firebase-admin==6.2.0

# Environment
python-dotenv==1.0.0

# For Step 2 (AI Integration)
google-generativeai==0.3.0
google-cloud-aiplatform==1.36.0
Pillow==10.0.1

# For Step 4 (Query Processing)
openai==1.3.0
langchain==0.0.350

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0